feat(03-01): decompose rate-limiter.ts into single-responsibility modules
Split 712-LOC monolith into focused modules:
- types.ts: type re-exports
- store.ts: EnhancedRateLimiter class
- analytics.ts, detection.ts: thin type re-exports
- configs.ts: preset configurations
- helpers.ts: standalone helper functions
- index.ts: singleton factory barrel export

Updated all consumer imports. All 133 tests pass.
