+++ src/app/api/send-email/action.ts
@@ -1,7 +1,7 @@
 import { headers } from 'next/headers'
 import { emailService } from '@/lib/email-service'
 import type { ContactFormData } from '@/types/api'
-import { getClientIdentifier } from '@/lib/rate-limiter'
+import { getClientIdentifier } from '@/lib/rate-limiter/helpers'
 import { createContextLogger } from '@/lib/logger';
 
 const logger = createContextLogger('SendemailAPI');
