+++ src/lib/rate-limiter/analytics.ts
@@ -1,0 +1,6 @@
+/**
+ * Rate Limiter Analytics
+ * Type re-exports for analytics. Analytics logic lives in store.ts (class methods using `this`).
+ */
+
+export type { RateLimitAnalytics } from '@/types/security'
