+++ src/lib/rate-limiter/detection.ts
@@ -1,0 +1,6 @@
+/**
+ * Rate Limiter Detection
+ * Type re-exports for detection. Detection logic lives in store.ts (class methods using `this`).
+ */
+
+export type { RateLimitResult } from '@/types/security'
